Finally fixed VC issue
It turns out that std::shared_ptr instances are COPIED rather
than passed by reference. This is done to be able to ref count
(duh!). So what happened is that when block.reset(target) is called
in check_vc(), it actually reset a *different* shared_ptr. In
other words, the original shared_ptr stored in the cache vector
is NOT updated. In effect, it is as if the valid bits are NOT
copied at all from VC to cache.

The fix I came up with is to just use raw pointers. First, I
allocate Blocks to the cache array. Then I simply pass them around
by reference through pointers.

Program now works for all test cases correctly.

I suck at C++.